Created Bus Class
The Bus class is created (initial commit) with the given attributes ID, bus route. The Bus is missing the function of calculating the fare. As it is not yet confirmed. It has all setters and getters. It has two constructors one for ID only, and the other one is for ID and bus route. The class throws an exception if there is any duplicate in IDs. The verification of duplicates is by measuring the size of a hash list which is static; it stores all of the IDs of the created bus. if the size changed then it is added, otherwise, it throws an exception. the function "number_of_available_seats" returns the number based on the number of seats valued '0' as it is a sign for the unbooked seats.  "Available_seats" is a function of the datatype of Vector<Integer>  which returns a Vector<Intger> with the number of available seats.
The built-in library used:
java.util.Set // for avoiding the duplicate as the set does not allow duplicate
java.util.TreeSet //for sorting the set 
java.util.Vector // for dynamically store data to the function "available seats"